requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.3
APScheduler==3.10.4
pysmb==1.2.9
